add todos os requisitos de moto e os testes da camada de controller

+++ src/controllers/GenericController.ts
@@ -17,17 +17,42 @@ enum ControllerErrors {
   badRequest = 'Bad request',
 }
 
-abstract class GenericController<T> {
-  abstract route: string;
+class GenericController<T> {
+  // route: string;
 
   protected errors = ControllerErrors;
   
-  constructor(protected service: GenericService<T>) {}
+  constructor(
+    public service: GenericService<T>,
+    public route: string, 
+  ) {}
 
-  abstract create(
+  get getRoute() { return this.route; }
+
+  // abstract create(
+  //   req: RequestWithBody<T>,
+  //   res: Response<T | ResponseError >,
+  // ): Promise<typeof res>;
+
+  create = async (
     req: RequestWithBody<T>,
-    res: Response<T | ResponseError >,
-  ): Promise<typeof res>;
+    res: Response<T | ResponseError>,
+  ): Promise<typeof res> => {
+    try {
+      const { body } = req;
+      const objs = await this.service.create(body);
+      if (!objs) {
+        return res.status(500).json({ error: this.errors.internal });
+      }
+      if ('error' in objs) {
+        return res.status(400)
+          .json({ error: objs.error.issues[0].message });
+      }
+      return res.status(201).json(objs);
+    } catch (err) {
+      return res.status(500).json({ error: this.errors.internal });
+    }
+  };
 
   read = async (
     _req: Request,
@@ -35,24 +60,19 @@ abstract class GenericController<T> {
   ): Promise<typeof res> => {
     try {
       const objs = await this.service.read();
-      return res.json(objs);
+      return res.status(200).json(objs);
     } catch (err) {
       return res.status(500).json({ error: this.errors.internal });
     }
   };
 
-  abstract readOne(
-    req: Request<{ id: string; }>,
-    res: Response<T | ResponseError>
-  ): Promise<typeof res>;
+  // abstract readOne(
+  //   req: Request<{ id: string; }>,
+  //   res: Response<T | ResponseError>
+  // ): Promise<typeof res>;
 
-  abstract update(
-    req: RequestWithBody<T>,
-    res: Response<T | ResponseError>,
-  ): Promise<typeof res>;
-
-  delete = async (
-    req: Request<{ id: string }>,
+  readOne = async (
+    req: Request,
     res: Response<T | ResponseError>,
   ): Promise<typeof res> => {
     const { id } = req.params;
@@ -61,9 +81,54 @@ abstract class GenericController<T> {
         .json({ error: 'Id must have 24 hexadecimal characters' });
     } 
     try {
-      const car = await this.service.delete(id);
-      return car 
-        ? res.status(204).json(car)
+      const obj = await this.service.readOne(id);
+      return obj 
+        ? res.json(obj)
+        : res.status(404).json({ error: this.errors.notFound });
+    } catch (error) {
+      return res.status(500).json({ error: this.errors.internal });
+    }
+  };
+
+  // abstract update(
+  //   req: RequestWithBody<T>,
+  //   res: Response<T | ResponseError>,
+  // ): Promise<typeof res>;
+
+  public update = async (
+    req: Request,
+    res: Response<T | ResponseError>,
+  ): Promise<typeof res> => {
+    try {
+      const obj = await this.service.update(req.params.id, req.body);
+
+      if (!obj) {
+        return res.status(404).json({ error: this.errors.notFound });
+      }
+      
+      if ('error' in obj) {
+        return res.status(400).json({ error: obj.error.issues[0].message });
+      }
+  
+      return res.status(200).json(obj);
+    } catch (error) {
+      return res.status(500).json({ error: this.errors.internal });
+    }
+  };
+
+  delete = async (
+    req: Request,
+    res: Response<T | ResponseError>,
+  ): Promise<typeof res> => {
+    const { id } = req.params;
+    if (id.length < 24) {
+      return res.status(400)
+        .json({ error: 'Id must have 24 hexadecimal characters' });
+    } 
+    try {
+      const obj = await this.service.delete(id);
+      return obj 
+        ? res.status(204).json(obj)
         : res.status(404).json({ error: this.errors.notFound });
     } catch (error) {
       return res.status(500).json({ error: this.errors.internal });
